backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation / settings
pydantic==2.7.1
pydantic-settings==2.2.1

# HTTP client (for calling ML service)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ML (import mode — installed separately in ml service for svc mode)
torch==2.3.0
transformers==4.41.0
scipy==1.13.0
numpy==1.26.4
